requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
authlib>=1.2.1
python-dotenv>=1.0.0
Flask-Cors>=4.0.0
requests>=2.31.0
SQLAlchemy>=2.0.22
Flask-SQLAlchemy>=3.1.1
pylint>=3.0.2
pytest>=7.4.3
coverage>=7.3.2